dom/html/test/forms/test_input_setting_value.html

<!DOCTYPE>
<html>
<head>
  <title>Test for setting input value</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<div id="display">
</div>
<div id="content"><input type="text"></div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(() => {
  let input = document.querySelector("input[type=text]");

  // Setting value during composition causes committing composition before setting the value.
  input.focus();
  let description = 'Setting input value at first "compositionupdate" event: ';
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "", `${description}input value should not have been modified at first "compositionupdate" event yet`);
    input.value = "def";
    is(input.value, "def", `${description}input value should be the specified value at "compositionupdate" event (after setting the value)`);
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    todo_is(input.value, "def", `${description}input value should be the specified value at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    todo_is(input.value, "def", `${description}input value should be the specified value at "input" event`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  todo_is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);

  input.value = "";
  description = 'Setting input value at second "compositionupdate" event: ';
  synthesizeCompositionChange(
    { "composition":
      { "string": "ab",
        "clauses":
        [
          { "length": 2, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 2, "length": 0 },
    });
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "ab", `${description}input value should not have been modified at second "compositionupdate" event yet`);
    input.value = "def";
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    is(input.value, "def", `${description}input value should be specified value at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    is(input.value, "def", `${description}input value should be specified value at "input" event`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);

  input.value = "";
  description = 'Setting input value at "input" event for first composition update: ';
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "", `${description}input value should not have been modified at first "compositionupdate" event yet`);
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    todo_is(input.value, "abc", `${description}input value should be the composition string at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    is(input.value, "abc", `${description}input value should be the composition string at "input" event`);
    input.value = "def";
    is(input.value, "def", `${description}input value should be the specified value at "input" event (after setting the value)`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);

  input.value = "";
  description = 'Setting input value at "input" event for second composition update: ';
  synthesizeCompositionChange(
    { "composition":
      { "string": "ab",
        "clauses":
        [
          { "length": 2, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 2, "length": 0 },
    });
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "ab", `${description}input value should not have been modified at second "compositionupdate" event yet`);
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    todo_is(input.value, "abc", `${description}input value should be the composition string at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    is(input.value, "abc", `${description}input value should be the composition string at "input" event`);
    input.value = "def";
    is(input.value, "def", `${description}input value should be the specified value at "input" event (after setting the value)`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);

  input.value = "";
  description = 'Setting input value and reframing at "input" event for first composition update: ';
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "", `${description}input value should not have been modified at first "compositionupdate" event yet`);
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    todo_is(input.value, "abc", `${description}input value should be the composition string at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    is(input.value, "abc", `${description}input value should be the composition string at "input" event`);
    input.value = "def";
    input.style.width = "1000px";
    is(input.value, "def", `${description}input value should be the specified value at "input" event (after setting the value)`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);
  input.style.width = "";

  input.value = "";
  description = 'Setting input value and reframing at "input" event for second composition update: ';
  synthesizeCompositionChange(
    { "composition":
      { "string": "ab",
        "clauses":
        [
          { "length": 2, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 2, "length": 0 },
    });
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "ab", `${description}input value should not have been modified at second "compositionupdate" event yet`);
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    todo_is(input.value, "abc", `${description}input value should be the composition string at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    is(input.value, "abc", `${description}input value should be the composition string at "input" event`);
    input.value = "def";
    input.style.width = "1000px";
    is(input.value, "def", `${description}input value should be the specified value at "input" event (after setting the value)`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);
  input.style.width = "";

  input.value = "";
  description = 'Setting input value and reframing with flushing layout at "input" event for first composition update: ';
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "", `${description}input value should not have been modified at first "compositionupdate" event yet`);
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    todo_is(input.value, "abc", `${description}input value should be the composition string at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    is(input.value, "abc", `${description}input value should be the composition string at "input" event`);
    input.value = "def";
    input.style.width = "1000px";
    document.documentElement.scrollTop;
    is(input.value, "def", `${description}input value should be the specified value at "input" event (after setting the value)`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);
  input.style.width = "";

  input.value = "";
  description = 'Setting input value and reframing with flushing layout at "input" event for second composition update: ';
  synthesizeCompositionChange(
    { "composition":
      { "string": "ab",
        "clauses":
        [
          { "length": 2, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 2, "length": 0 },
    });
  input.addEventListener("compositionupdate", (aEvent) => {
    is(input.value, "ab", `${description}input value should not have been modified at second "compositionupdate" event yet`);
  }, {once: true});
  input.addEventListener("compositionend", (aEvent) => {
    todo_is(input.value, "abc", `${description}input value should be the composition string at "compositionend" event`);
  }, {once: true});
  input.addEventListener("input", (aEvent) => {
    is(input.value, "abc", `${description}input value should be the composition string at "input" event`);
    input.value = "def";
    input.style.width = "1000px";
    document.documentElement.scrollTop;
    is(input.value, "def", `${description}input value should be the specified value at "input" event (after setting the value)`);
  }, {once: true});
  synthesizeCompositionChange(
    { "composition":
      { "string": "abc",
        "clauses":
        [
          { "length": 3, "attr": COMPOSITION_ATTR_RAW_CLAUSE }
        ]
      },
      "caret": { "start": 3, "length": 0 },
    });
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should be set to specified value after the last "input" event`);
  input.style.width = "";

  // autocomplete and correcting misspelled word by spellchecker cause an "input" event with same path as setting input value.
  input.value = "";
  description = 'Setting input value at "input" event whose inputType is "insertReplacementText';
  let inputEventFired = false;
  input.addEventListener("input", (aEvent) => {
    is(aEvent.inputType, "insertReplacementText", `${description}inputType of "input" event should be "insertReplacementText"`);
    inputEventFired = true;
    is(input.value, "abc", `${description}input value should be inserted value at "input" event (before setting value)`);
    input.value = "def";
    is(input.value, "def", `${description}input value should be specified value at "input" event (after setting value)`);
  }, {once: true});
  SpecialPowers.wrap(input).setUserInput("abc");
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should keep the specified value after the last "input" event`);
  ok(inputEventFired, `${description}"input" event should've been fired for setUserInput("abc")`);

  input.value = "";
  description = 'Setting input value and reframing at "input" event whose inputType is "insertReplacementText';
  inputEventFired = false;
  input.addEventListener("input", (aEvent) => {
    is(aEvent.inputType, "insertReplacementText", `${description}inputType of "input" event should be "insertReplacementText"`);
    inputEventFired = true;
    is(input.value, "abc", `${description}input value should be inserted value at "input" event (before setting value)`);
    input.value = "def";
    input.style.width = "1000px";
    is(input.value, "def", `${description}input value should be specified value at "input" event (after setting value)`);
  }, {once: true});
  SpecialPowers.wrap(input).setUserInput("abc");
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should keep the specified value after the last "input" event`);
  ok(inputEventFired, `${description}"input" event should've been fired for setUserInput("abc")`);
  input.style.width = "";

  input.value = "";
  description = 'Setting input value and reframing with flushing layout at "input" event whose inputType is "insertReplacementText';
  inputEventFired = false;
  input.addEventListener("input", (aEvent) => {
    is(aEvent.inputType, "insertReplacementText", `${description}inputType of "input" event should be "insertReplacementText"`);
    inputEventFired = true;
    is(input.value, "abc", `${description}input value should be inserted value at "input" event (before setting value)`);
    input.value = "def";
    input.style.width = "1000px";
    document.documentElement.scrollTop;
    is(input.value, "def", `${description}input value should be specified value at "input" event (after setting value)`);
  }, {once: true});
  SpecialPowers.wrap(input).setUserInput("abc");
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should keep the specified value after the last "input" event`);
  ok(inputEventFired, `${description}"input" event should've been fired for setUserInput("abc")`);
  input.style.width = "";

  input.value = "";
  description = 'Setting input value and destroying the frame at "input" event whose inputType is "insertReplacementText';
  inputEventFired = false;
  input.addEventListener("input", (aEvent) => {
    is(aEvent.inputType, "insertReplacementText", `${description}inputType of "input" event should be "insertReplacementText"`);
    inputEventFired = true;
    is(input.value, "abc", `${description}input value should be inserted value at "input" event (before setting value)`);
    input.value = "def";
    input.style.display = "none";
    is(input.value, "def", `${description}input value should be specified value at "input" event (after setting value)`);
  }, {once: true});
  SpecialPowers.wrap(input).setUserInput("abc");
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "def", `${description}input value should keep the specified value after the last "input" event`);
  ok(inputEventFired, `${description}"input" event should've been fired for setUserInput("abc")`);
  input.style.display = "inline";

  input.value = "";
  description = 'Changing input type at "input" event whose inputType is "insertReplacementText';
  inputEventFired = false;
  input.addEventListener("input", (aEvent) => {
    is(aEvent.inputType, "insertReplacementText", `${description}inputType of "input" event should be "insertReplacementText"`);
    inputEventFired = true;
    is(input.value, "abc", `${description}input value should be inserted value at "input" event (before changing type)`);
    input.type = "button";
    is(input.value, "abc", `${description}input value should keep inserted value at "input" event (after changing type)`);
  }, {once: true});
  SpecialPowers.wrap(input).setUserInput("abc");
  is(input.value, "abc", `${description}input value should keep inserted value after the last "input" event`);
  is(input.type, "button", `${description}input type should be changed correctly`);
  ok(inputEventFired, `${description}"input" event should've been fired for setUserInput("abc")`);
  input.type = "text";
  is(input.value, "abc", `${description}input value should keep inserted value immediately after restoring the type`);
  todo(SpecialPowers.wrap(input).editor !== null, `${description}restoring input type should create editor if it's focused element`);
  input.blur();
  input.focus();
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "abc", `${description}input value should keep inserted value after creating editor`);

  input.value = "";
  description = 'Changing input type and flush layout at "input" event whose inputType is "insertReplacementText';
  inputEventFired = false;
  input.addEventListener("input", (aEvent) => {
    is(aEvent.inputType, "insertReplacementText", `${description}inputType of "input" event should be "insertReplacementText"`);
    inputEventFired = true;
    is(input.value, "abc", `${description}input value should be inserted value at "input" event (before changing type)`);
    input.type = "button";
    input.getBoundingClientRect().height;
    is(input.value, "abc", `${description}input value should keep inserted value at "input" event (after changing type)`);
  }, {once: true});
  SpecialPowers.wrap(input).setUserInput("abc");
  is(input.value, "abc", `${description}input value should keep inserted value after the last "input" event`);
  is(input.type, "button", `${description}input type should be changed correctly`);
  ok(inputEventFired, `${description}"input" event should've been fired for setUserInput("abc")`);
  input.type = "text";
  is(input.value, "abc", `${description}input value should keep inserted value immediately after restoring the type`);
  todo(SpecialPowers.wrap(input).editor !== null, `${description}restoring input type should create editor if it's focused element`);
  input.blur();
  input.focus();
  is(SpecialPowers.wrap(input).editor.rootElement.firstChild.wholeText, input.value,
     `${description}native anonymous text node should have exactly same value as value of <input> element`);
  is(input.value, "abc", `${description}input value should keep inserted value after creating editor`);

  SimpleTest.finish();
});
</script>
</body>
</html>
